+++ lib/odoo.bash
@@ -11,6 +11,8 @@ fi
 ohelper_require 'install';
 ohelper_require 'server';
 ohelper_require 'fetch';
+ohelper_require 'git';
+ohelper_require 'scaffold';
 # ----------------------------------------------------------------------------------------
 
 
@@ -94,9 +96,5 @@ function odoo_update_sources {
 
 # odoo_scaffold <addon_name> [addon_path]
 function odoo_scaffold {
-    local addon_name=$1;
-    local addon_path=${2:-$REPOSITORIES_DIR};
-
-    odoo_py scaffold $addon_name $addon_path;
-    link_module $addon_path/$addon_name;
+    scaffold_default "$@";
 }